streamlit_app_v.0.4/requirements.txt
streamlit
plotly
Pillow
pandas
numpy
scipy
streamlit_plotly_events
sqlalchemy
psycopg2-binary
streamlit-aggrid
